R! FragId became ushort instead of byte

+++ Pixockets/FragmentedPacket.cs
@@ -4,7 +4,7 @@ namespace Pixockets
 {
     public class FragmentedPacket
     {
-        public byte FragId;
+        public ushort FragId;
         public List<FragmentBuffer> Buffers = new List<FragmentBuffer>();
         public int LastActive;
     }
